Project 9 final FINAL commit after war file submission during extended late period with no penalty

+++ server/src/main/java/business/cart/ShoppingCartItem.java
@@ -44,7 +44,7 @@ public class ShoppingCartItem {
 	}
 
 	@JsonIgnore
-	public long getBookPrice() {
+	public int getBookPrice() {
 		return bookForm.getPrice();
 	}
